requirements.txt
openai-whisper
torch
transformers
numpy
ffmpeg-python
librosa
scikit-learn
pandas
datasets
matplotlib
seaborn
fastapi
uvicorn
pydantic
streamlit 
requests
python-multipart


